+++ daemon/dataSourcePuller.py
@@ -1,0 +1,4 @@
+
+
+class DataSourcePuller:
+    pass
