+++ daemon/dataSources.py
@@ -1,3 +1,3 @@
-from .dataSourceModel import dataSourceModel
+from .helpers import create_notification
 from shared.models import Notification
 from datetime import datetime
